refactor: move Shop Logo directory
